Fixed warnings

+++ Implementierung/src/Implementierung/simd_sobel.h
@@ -2,8 +2,8 @@
 #include <stddef.h>
 #include <immintrin.h>
 
-#define COMP_255 255, 255, 255, 255, 255, 255, 255, 255
-#define ZERO_EVEN_BYTES_MASK 0xFF00, 0xFF00, 0xFF00, 0xFF00, 0xFF00, 0xFF00, 0xFF00, 0xFF00
+#define COMP_255 (uint16_t)255, (uint16_t)255, (uint16_t)255, (uint16_t)255, (uint16_t)255, (uint16_t)255, (uint16_t)255, (uint16_t)255
+#define ZERO_EVEN_BYTES_MASK (uint16_t)0xFF00, (uint16_t)0xFF00, (uint16_t)0xFF00, (uint16_t)0xFF00, (uint16_t)0xFF00, (uint16_t)0xFF00, (uint16_t)0xFF00, (uint16_t)0xFF00
 
 void simd_sobel(uint8_t* img_in, size_t width, size_t height, uint8_t* img_out);
